Move old code into _app
