+++ _app/routes/auth.callback.dropbox.ts
@@ -1,8 +1,7 @@
 import type { LoaderArgs } from '@remix-run/node';
 import { redirect } from '@remix-run/node';
-
-import { config } from '~/config';
-import { DropboxClient, storage } from '~/services/dropbox.server';
+import { config } from '_app/config';
+import { DropboxClient, storage } from '_app/services/dropbox.server';
 
 export async function loader({ request }: LoaderArgs) {
   let [dbx, session] = await DropboxClient.fromRequest(request);
